server_def move to common.h
