+++ src/local.h
@@ -27,44 +27,10 @@
 #include <libcork/ds.h>
 
 #include "encrypt.h"
-#include "obfs/obfs.h"
 #include "jconf.h"
 #include "protocol.h"
 
 #include "common.h"
-
-typedef struct server_def {
-    char *host;
-    int port;
-    int udp_port;
-    struct sockaddr_storage *addr; // resolved address
-    struct sockaddr_storage *addr_udp; // resolved address
-    int addr_len;
-    int addr_udp_len;
-
-    char *psw; // raw password
-    cipher_env_t cipher;
-
-    struct cork_dllist connections;
-
-    // SSR
-    char *protocol_name; // for logging use only?
-    char *obfs_name; // for logging use only?
-
-    char *protocol_param;
-    char *obfs_param;
-
-    obfs_class *protocol_plugin;
-    obfs_class *obfs_plugin;
-
-    void *protocol_global;
-    void *obfs_global;
-
-    int enable;
-    char *id;
-    char *group;
-    int udp_over_tcp;
-} server_def_t;
 
 // use this as a profile or environment
 typedef struct listen_ctx{
